LayoutTests/canvas/philip/tests/size.attributes.idl-expected.txt

< [index] >
size.attributes.idl
Getting/setting width/height IDL attributes
References: size.width size.height
Actual output:
Passed
